+++ src/NATS.Client.Core/Commands/CommandWriter.cs
@@ -1,7 +1,6 @@
 using System.Buffers;
 using System.IO.Pipelines;
 using System.Net.Sockets;
-using System.Numerics;
 using System.Runtime.CompilerServices;
 using System.Threading.Channels;
 using Microsoft.Extensions.Logging;
@@ -35,7 +34,7 @@ internal sealed class CommandWriter : IAsyncDisposable
     private readonly PipeReader _pipeReader;
     private readonly PipeWriter _pipeWriter;
     private ISocketConnection? _socketConnection;
-    private bool _disposed;
+    private volatile bool _disposed;
 
     public CommandWriter(ObjectPool pool, NatsOpts opts, ConnectionStatsCounter counter, Action<PingCommand> enqueuePing, TimeSpan? overrideCommandTimeout = default)
     {
@@ -68,38 +67,26 @@ internal sealed class CommandWriter : IAsyncDisposable
 
     public async ValueTask DisposeAsync()
     {
+        if (_disposed)
+        {
+            return;
+        }
+
+        _disposed = true;
+
 #if NET6_0
         _cts.Cancel();
 #else
         await _cts.CancelAsync().ConfigureAwait(false);
 #endif
 
-        while (!_channelLock.Writer.TryWrite(1))
-        {
-            await _channelLock.Writer.WaitToWriteAsync().ConfigureAwait(false);
-        }
+        await _channelLock.Writer.WriteAsync(1).ConfigureAwait(false);
+        _channelLock.Writer.TryComplete();
 
-        try
-        {
-            if (_disposed)
-            {
-                return;
-            }
+        await _pipeWriter.CompleteAsync().ConfigureAwait(false);
+        await _pipeReader.CompleteAsync().ConfigureAwait(false);
 
-            _disposed = true;
-
-            await _pipeWriter.CompleteAsync().ConfigureAwait(false);
-            await _pipeReader.CompleteAsync().ConfigureAwait(false);
-
-            await _readerLoopTask.ConfigureAwait(false);
-        }
-        finally
-        {
-            while (!_channelLock.Reader.TryRead(out _))
-            {
-                await _channelLock.Reader.WaitToReadAsync().ConfigureAwait(false);
-            }
-        }
+        await _readerLoopTask.ConfigureAwait(false);
     }
 
     public async ValueTask ConnectAsync(ClientOpts connectOpts, CancellationToken cancellationToken)
@@ -142,21 +129,10 @@ internal sealed class CommandWriter : IAsyncDisposable
 
     public async ValueTask PingAsync(PingCommand pingCommand, CancellationToken cancellationToken)
     {
-        Interlocked.Increment(ref _counter.PendingMessages);
-
-        try
-        {
-            await _channelLock.Writer.WriteAsync(1, cancellationToken).ConfigureAwait(false);
-        }
-        catch (OperationCanceledException)
+        if (!await LockAsync(cancellationToken).ConfigureAwait(false))
         {
             return;
         }
-        catch (ChannelClosedException)
-        {
-            return;
-        }
-
 
         try
         {
@@ -172,32 +148,16 @@ internal sealed class CommandWriter : IAsyncDisposable
         }
         finally
         {
-            while (!_channelLock.Reader.TryRead(out _))
-            {
-                await _channelLock.Reader.WaitToReadAsync(cancellationToken).ConfigureAwait(false);
-            }
-
-            Interlocked.Decrement(ref _counter.PendingMessages);
+            await UnLockAsync(cancellationToken).ConfigureAwait(true);
         }
     }
 
     public async ValueTask PongAsync(CancellationToken cancellationToken = default)
     {
-        Interlocked.Increment(ref _counter.PendingMessages);
-
-        try
-        {
-            await _channelLock.Writer.WriteAsync(1, cancellationToken).ConfigureAwait(false);
-        }
-        catch (OperationCanceledException)
+        if (!await LockAsync(cancellationToken).ConfigureAwait(false))
         {
             return;
         }
-        catch (ChannelClosedException)
-        {
-            return;
-        }
-
 
         try
         {
@@ -211,12 +171,7 @@ internal sealed class CommandWriter : IAsyncDisposable
         }
         finally
         {
-            while (!_channelLock.Reader.TryRead(out _))
-            {
-                await _channelLock.Reader.WaitToReadAsync(cancellationToken).ConfigureAwait(false);
-            }
-
-            Interlocked.Decrement(ref _counter.PendingMessages);
+            await UnLockAsync(cancellationToken).ConfigureAwait(true);
         }
     }
 
@@ -241,21 +196,10 @@ internal sealed class CommandWriter : IAsyncDisposable
 
     public async ValueTask SubscribeAsync(int sid, string subject, string? queueGroup, int? maxMsgs, CancellationToken cancellationToken)
     {
-        Interlocked.Increment(ref _counter.PendingMessages);
-
-        try
-        {
-            await _channelLock.Writer.WriteAsync(1, cancellationToken).ConfigureAwait(false);
-        }
-        catch (OperationCanceledException)
+        if (!await LockAsync(cancellationToken).ConfigureAwait(false))
         {
             return;
         }
-        catch (ChannelClosedException)
-        {
-            return;
-        }
-
 
         try
         {
@@ -269,28 +213,13 @@ internal sealed class CommandWriter : IAsyncDisposable
         }
         finally
         {
-            while (!_channelLock.Reader.TryRead(out _))
-            {
-                await _channelLock.Reader.WaitToReadAsync(cancellationToken).ConfigureAwait(false);
-            }
-
-            Interlocked.Decrement(ref _counter.PendingMessages);
+            await UnLockAsync(cancellationToken).ConfigureAwait(true);
         }
     }
 
     public async ValueTask UnsubscribeAsync(int sid, int? maxMsgs, CancellationToken cancellationToken)
     {
-        Interlocked.Increment(ref _counter.PendingMessages);
-
-        try
-        {
-            await _channelLock.Writer.WriteAsync(1, cancellationToken).ConfigureAwait(false);
-        }
-        catch (OperationCanceledException)
-        {
-            return;
-        }
-        catch (ChannelClosedException)
+        if (!await LockAsync(cancellationToken).ConfigureAwait(false))
         {
             return;
         }
@@ -307,36 +236,17 @@ internal sealed class CommandWriter : IAsyncDisposable
         }
         finally
         {
-            while (!_channelLock.Reader.TryRead(out _))
-            {
-                await _channelLock.Reader.WaitToReadAsync(cancellationToken).ConfigureAwait(false);
-            }
-
-            Interlocked.Decrement(ref _counter.PendingMessages);
+            await UnLockAsync(cancellationToken).ConfigureAwait(true);
         }
     }
-
-    [MethodImpl(MethodImplOptions.NoInlining)]
-    private static void ThrowOnDisconnected() => throw new NatsException("Connection hasn't been established yet.");
 
     [AsyncMethodBuilder(typeof(PoolingAsyncValueTaskMethodBuilder))]
     private async ValueTask PublishLockedAsync(string subject, string? replyTo,  NatsPooledBufferWriter<byte> payloadBuffer, NatsPooledBufferWriter<byte>? headersBuffer, CancellationToken cancellationToken)
     {
-        Interlocked.Increment(ref _counter.PendingMessages);
-
-        try
-        {
-            await _channelLock.Writer.WriteAsync(1, cancellationToken).ConfigureAwait(false);
-        }
-        catch (OperationCanceledException)
+        if (!await LockAsync(cancellationToken).ConfigureAwait(false))
         {
             return;
         }
-        catch (ChannelClosedException)
-        {
-            return;
-        }
-
 
         try
         {
@@ -363,13 +273,45 @@ internal sealed class CommandWriter : IAsyncDisposable
         }
         finally
         {
-            while (!_channelLock.Reader.TryRead(out _))
+            await UnLockAsync(cancellationToken).ConfigureAwait(true);
+        }
+    }
+
+    private async Task UnLockAsync(CancellationToken cancellationToken)
+    {
+        while (!_channelLock.Reader.TryRead(out _))
+        {
+            try
             {
                 await _channelLock.Reader.WaitToReadAsync(cancellationToken).ConfigureAwait(false);
             }
-
-            Interlocked.Decrement(ref _counter.PendingMessages);
+            catch (OperationCanceledException)
+            {
+                break;
+            }
         }
+
+        Interlocked.Decrement(ref _counter.PendingMessages);
+    }
+
+    private async Task<bool> LockAsync(CancellationToken cancellationToken)
+    {
+        Interlocked.Increment(ref _counter.PendingMessages);
+
+        try
+        {
+            await _channelLock.Writer.WriteAsync(1, cancellationToken).ConfigureAwait(false);
+        }
+        catch (OperationCanceledException)
+        {
+            return false;
+        }
+        catch (ChannelClosedException)
+        {
+            return false;
+        }
+
+        return true;
     }
 
     private async Task ReaderLoopAsync()
@@ -448,215 +390,5 @@ internal sealed class CommandWriter : IAsyncDisposable
                 _logger.LogError(NatsLogEvents.Buffer, e, "Unexpected error in send buffer reader loop");
             }
         }
-    }
-}
-
-internal sealed class PriorityCommandWriter : IAsyncDisposable
-{
-    private int _disposed;
-
-    public PriorityCommandWriter(ObjectPool pool, ISocketConnection socketConnection, NatsOpts opts, ConnectionStatsCounter counter, Action<PingCommand> enqueuePing)
-    {
-        CommandWriter = new CommandWriter(pool, opts, counter, enqueuePing, overrideCommandTimeout: TimeSpan.MaxValue);
-        CommandWriter.Reset(socketConnection);
-    }
-
-    public CommandWriter CommandWriter { get; }
-
-    public async ValueTask DisposeAsync()
-    {
-        if (Interlocked.Increment(ref _disposed) == 1)
-        {
-            // disposing command writer marks pipe writer as complete
-            await CommandWriter.DisposeAsync().ConfigureAwait(false);
-        }
-    }
-}
-
-internal sealed class NatsPooledBufferWriter<T> : IBufferWriter<T>, IObjectPoolNode<NatsPooledBufferWriter<T>>
-{
-    private const int DefaultInitialBufferSize = 256;
-
-    private readonly ArrayPool<T> _pool;
-    private T[]? _array;
-    private int _index;
-    private NatsPooledBufferWriter<T>? _next;
-
-    public NatsPooledBufferWriter()
-    {
-        _pool = ArrayPool<T>.Shared;
-        _array = _pool.Rent(DefaultInitialBufferSize);
-        _index = 0;
-    }
-
-    public ref NatsPooledBufferWriter<T>? NextNode => ref _next;
-
-    /// <summary>
-    /// Gets the data written to the underlying buffer so far, as a <see cref="ReadOnlyMemory{T}"/>.
-    /// </summary>
-    public ReadOnlyMemory<T> WrittenMemory
-    {
-        get
-        {
-            var array = _array;
-
-            if (array is null)
-            {
-                ThrowObjectDisposedException();
-            }
-
-            return array!.AsMemory(0, _index);
-        }
-    }
-
-    /// <summary>
-    /// Gets the data written to the underlying buffer so far, as a <see cref="ReadOnlySpan{T}"/>.
-    /// </summary>
-    public ReadOnlySpan<T> WrittenSpan
-    {
-        get
-        {
-            var array = _array;
-
-            if (array is null)
-            {
-                ThrowObjectDisposedException();
-            }
-
-            return array!.AsSpan(0, _index);
-        }
-    }
-
-    /// <summary>
-    /// Gets the amount of data written to the underlying buffer so far.
-    /// </summary>
-    public int WrittenCount
-    {
-        get => _index;
-    }
-
-    /// <inheritdoc/>
-    public void Advance(int count)
-    {
-        var array = _array;
-
-        if (array is null)
-        {
-            ThrowObjectDisposedException();
-        }
-
-        if (count < 0)
-        {
-            ThrowArgumentOutOfRangeExceptionForNegativeCount();
-        }
-
-        if (_index > array!.Length - count)
-        {
-            ThrowArgumentExceptionForAdvancedTooFar();
-        }
-
-        _index += count;
-    }
-
-    /// <inheritdoc/>
-    public Memory<T> GetMemory(int sizeHint = 0)
-    {
-        CheckBufferAndEnsureCapacity(sizeHint);
-
-        return _array.AsMemory(_index);
-    }
-
-    /// <inheritdoc/>
-    public Span<T> GetSpan(int sizeHint = 0)
-    {
-        CheckBufferAndEnsureCapacity(sizeHint);
-
-        return _array.AsSpan(_index);
-    }
-
-    public void Reset()
-    {
-        if (_array != null)
-            _pool.Return(_array);
-        _array = _pool.Rent(DefaultInitialBufferSize);
-        _index = 0;
-    }
-
-    /// <inheritdoc/>
-    public override string ToString()
-    {
-        // See comments in MemoryOwner<T> about this
-        if (typeof(T) == typeof(char) &&
-            _array is char[] chars)
-        {
-            return new(chars, 0, _index);
-        }
-
-        // Same representation used in Span<T>
-        return $"NatsPooledBufferWriter<{typeof(T)}>[{_index}]";
-    }
-
-    [MethodImpl(MethodImplOptions.NoInlining)]
-    private static void ThrowArgumentOutOfRangeExceptionForNegativeCount() => throw new ArgumentOutOfRangeException("count", "The count can't be a negative value.");
-
-    [MethodImpl(MethodImplOptions.NoInlining)]
-    private static void ThrowArgumentOutOfRangeExceptionForNegativeSizeHint() => throw new ArgumentOutOfRangeException("sizeHint", "The size hint can't be a negative value.");
-
-    [MethodImpl(MethodImplOptions.NoInlining)]
-    private static void ThrowArgumentExceptionForAdvancedTooFar() => throw new ArgumentException("The buffer writer has advanced too far.");
-
-    [MethodImpl(MethodImplOptions.NoInlining)]
-    private static void ThrowObjectDisposedException() => throw new ObjectDisposedException("The current buffer has already been disposed.");
-
-    /// <summary>
-    /// Ensures that <see cref="_array"/> has enough free space to contain a given number of new items.
-    /// </summary>
-    /// <param name="sizeHint">The minimum number of items to ensure space for in <see cref="_array"/>.</param>
-    [MethodImpl(MethodImplOptions.AggressiveInlining)]
-    private void CheckBufferAndEnsureCapacity(int sizeHint)
-    {
-        var array = _array;
-
-        if (array is null)
-        {
-            ThrowObjectDisposedException();
-        }
-
-        if (sizeHint < 0)
-        {
-            ThrowArgumentOutOfRangeExceptionForNegativeSizeHint();
-        }
-
-        if (sizeHint == 0)
-        {
-            sizeHint = 1;
-        }
-
-        if (sizeHint > array!.Length - _index)
-        {
-            ResizeBuffer(sizeHint);
-        }
-    }
-
-    /// <summary>
-    /// Resizes <see cref="_array"/> to ensure it can fit the specified number of new items.
-    /// </summary>
-    /// <param name="sizeHint">The minimum number of items to ensure space for in <see cref="_array"/>.</param>
-    [MethodImpl(MethodImplOptions.NoInlining)]
-    private void ResizeBuffer(int sizeHint)
-    {
-        var minimumSize = (uint)_index + (uint)sizeHint;
-
-        // The ArrayPool<T> class has a maximum threshold of 1024 * 1024 for the maximum length of
-        // pooled arrays, and once this is exceeded it will just allocate a new array every time
-        // of exactly the requested size. In that case, we manually round up the requested size to
-        // the nearest power of two, to ensure that repeated consecutive writes when the array in
-        // use is bigger than that threshold don't end up causing a resize every single time.
-        if (minimumSize > 1024 * 1024)
-        {
-            minimumSize = BitOperations.RoundUpToPowerOf2(minimumSize);
-        }
-
-        _pool.Resize(ref _array, (int)minimumSize);
     }
 }
